+++ imdb_project/src/components/pages/Home.tsx
@@ -1,3 +1,4 @@
+
 const Home = () => {
     return ( 
        <>
